+++ src/vue/pages/council/tabs/CouncilOverviewTab.vue
@@ -1,22 +1,31 @@
 <template>
   <div class="council-overview-tab">
-    <council-summary />
-
     <div class="council-overview-tab__topbar">
-      <app-button
-        scheme="primary"
-        :text="$t('council-page.council-overview-tab.vote-btn')"
-        disabled
+      <council-summary
+        class="council-overview-tab__summary"
+        :elections="electionsInfo"
+        :has-elections="hasElections"
       />
-      <app-button
-        scheme="primary"
-        :text="$t('council-page.council-overview-tab.submit-candidacy-btn')"
-        disabled
-      />
+
+      <div class="council-overview-tab__actions">
+        <app-button
+          class="council-overview-tab__button"
+          scheme="primary"
+          :text="$t('council-page.council-overview-tab.vote-btn')"
+          disabled
+        />
+        <app-button
+          class="council-overview-tab__button"
+          scheme="primary"
+          :text="$t('council-page.council-overview-tab.submit-candidacy-btn')"
+          disabled
+        />
+      </div>
     </div>
 
-    <div class="council-overview-tab__body">
+    <div :class="{'council-overview-tab__body': hasElections}">
       <councils-list
+        v-if="hasElections"
         :elections="electionsInfo"
         :prime="prime"
         :all-votes="allVotes"
@@ -31,6 +40,7 @@
         class="council-overview-tab__list"
       />
       <councils-list
+        v-if="hasElections"
         :elections="electionsInfo"
         :prime="prime"
         :all-votes="allVotes"
@@ -62,13 +72,13 @@ export default {
     const prime = useCall(api.derive.council.prime)
     const electionsInfo = useCall(api.derive.elections.info)
     const allVotes = useAllVotes()
-    const moduleElections = useModuleElections()
+    const { hasElections } = useModuleElections()
 
     return {
       prime,
       electionsInfo,
       allVotes,
-      moduleElections,
+      hasElections,
     }
   },
 }
@@ -82,9 +92,27 @@ export default {
 
 .council-overview-tab__topbar {
   display: flex;
-  justify-content: flex-end;
-  gap: 2rem;
+  align-items: center;
+  justify-content: space-between;
   margin: 3rem 0;
+  gap: 2rem;
+  flex-wrap: wrap;
+
+  @include respond-to($tablet) {
+    flex-direction: column;
+    align-items: flex-end;
+    width: 100%;
+  }
+}
+
+.council-overview-tab__summary {
+  @include respond-to($tablet) {
+    width: 100%;
+  }
+}
+
+.council-overview-tab__actions {
+  display: flex;
 
   @include respond-to($x-small) {
     flex-direction: column;
@@ -92,10 +120,21 @@ export default {
   }
 }
 
+.council-overview-tab__button {
+  & + & {
+    margin-left: 2rem;
+
+    @include respond-to($x-small) {
+      margin-left: 0;
+      margin-top: 2rem;
+    }
+  }
+}
+
 .council-overview-tab__body {
   display: grid;
   gap: 2rem 3.4rem;
-  grid-template-columns: 0.3fr 0.7fr;
+  grid-template-columns: minmax(15rem, 0.3fr)  minmax(15rem, 0.7fr);
 
   @include respond-to($sidebar-hide-bp) {
     grid-template-columns: 1fr;