+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 - Gilt queue list
 - Gilt summary
+- Council lists
+- Council summary
 
 ## [0.1.0] - 2021-11-26
 #### Added